Refactor database migrations and seeders
- Removed obsolete migration files for OAuth tokens, OAuth scopes, password resets, roles, permissions, user roles, role permissions, and user permissions.
- Updated migration registration in `database_migrate.go` to reflect new migration structure.
- Created a new `DatabaseSeeder` to manage seeding operations, replacing the previous `UsersSeeder`.
- Introduced individual seeders for permissions, roles, and users to streamline the seeding process.
- Implemented role assignment logic within the `UserSeeder`.
- Removed the old `RolesPermissionsSeeder` in favor of a more modular approach.
- Added error handling and logging for better traceability during seeding.

+++ app/http/controllers/api/v1/auth/login_controller.go
@@ -49,7 +49,7 @@ func Login(context *gin.Context) {
 		return
 	}
 	client := clients[0]
-	token, err := oauth_models.CreateToken(int64(storedUser.ID), client.ID, "")
+	token, err := oauth_models.CreateToken(storedUser.ID, client.ID, "")
 	if err != nil {
 		context.JSON(http.StatusInternalServerError, gin.H{"errors": []gin.H{{
 			"status": "500",
@@ -59,7 +59,7 @@ func Login(context *gin.Context) {
 		return
 	}
 
-	resource := resources.NewAuthResource(uint(storedUser.ID), storedUser.FirstName+" "+storedUser.LastName, storedUser.Email, token.AccessToken)
+	resource := resources.NewAuthResource(storedUser.ID, storedUser.FirstName+" "+storedUser.LastName, storedUser.Email, token.AccessToken)
 	response := resources.NewAuthLoginResponse(resource, token.RefreshToken, 86400, token.Scopes)
 	context.JSON(http.StatusOK, response)
 }
