+++ app/http/controllers/api/v1/auth/login_controller_new.go
@@ -74,7 +74,7 @@ func LoginNew(context *gin.Context) {
 	}
 
 	client := clients[0]
-	token, err := oauth_models.CreateToken(int64(storedUser.ID), client.ID, "")
+	token, err := oauth_models.CreateToken(storedUser.ID, client.ID, "")
 	if err != nil {
 		context.JSON(http.StatusInternalServerError, validators.ValidationResponse{
 			Errors: []validators.ValidationErrorResponse{{
@@ -91,7 +91,7 @@ func LoginNew(context *gin.Context) {
 	}
 
 	// Respuesta exitosa
-	resource := resources.NewAuthResource(uint(storedUser.ID), storedUser.FirstName+" "+storedUser.LastName, storedUser.Email, token.AccessToken)
+	resource := resources.NewAuthResource(storedUser.ID, storedUser.FirstName+" "+storedUser.LastName, storedUser.Email, token.AccessToken)
 	response := resources.NewAuthLoginResponse(resource, token.RefreshToken, 86400, token.Scopes)
 	context.JSON(http.StatusOK, response)
 }
